Fix web dashboard ESP32 build

+++ src/web_manager.cpp
@@ -4,6 +4,7 @@
 #include <WiFi.h>
 #include <math.h>
 
+#include "board_config.h"
 #include "config.h"
 #include "davis_radio.h"
 #include "meteobridge_client.h"
@@ -21,8 +22,8 @@ String esc(const String &s){
   for(size_t i=0;i<s.length();i++){ char c=s[i]; if(c=='&')o+=F("&amp;"); else if(c=='<')o+=F("&lt;"); else if(c=='>')o+=F("&gt;"); else if(c=='\"')o+=F("&quot;"); else o+=c; }
   return o;
 }
-String jf(float v,uint8_t d=1){ return isfinite(v)?String(v,d):String("null"); }
-String hv(float v,uint8_t d=1,const char *unit=""){ if(!isfinite(v))return "--"; return String(v,d)+unit; }
+String jf(float v,unsigned int d=1U){ return isfinite(v)?String(v,d):String("null"); }
+String hv(float v,unsigned int d=1U,const char *unit=""){ if(!isfinite(v))return "--"; return String(v,d)+unit; }
 
 String page(){
   const StationState &s=*statePtr;
@@ -32,9 +33,9 @@ String page(){
   h+=F("<!doctype html><html><head><meta name='viewport' content='width=device-width,initial-scale=1'><meta http-equiv='refresh' content='15'><title>Davis Gateway</title><style>body{font-family:Arial,sans-serif;max-width:980px;margin:22px auto;padding:0 14px;color:#222}.grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(180px,1fr));gap:10px}.c{border:1px solid #ccc;border-radius:8px;padding:12px}.v{font-size:1.5em;font-weight:700}a,button{display:inline-block;margin:5px 5px 5px 0;padding:9px 12px}table{border-collapse:collapse;width:100%}td{padding:6px;border-bottom:1px solid #eee}.ok{color:#087b32}.ko{color:#a22}</style></head><body>");
   h+=F("<h1>ESP32 Davis Weather Gateway</h1><p>"); h+=BOARD_NAME; h+=F(" - firmware "); h+=FIRMWARE_VERSION; h+=F("</p><div class='grid'>");
   auto card=[&](const char *name,const String &value){ h+=F("<div class='c'><b>");h+=name;h+=F("</b><div class='v'>");h+=value;h+=F("</div></div>");};
-  card("Temperatura",hv(s.outTempC,1," C")); card("Umidita",hv(s.outHumidity,0," %")); card("Vento",hv(s.windKmh,1," km/h")); card("Raffica 10m",hv(s.windGustKmh,1," km/h"));
-  card("Direzione",hv(s.windDirDeg,0," deg")); card("Pioggia oggi",hv(s.rainDayMm,1," mm")); card("Rain rate",hv(s.rainRateMmH,1," mm/h")); card("Pressione",hv(s.pressureHpa,1," hPa"));
-  card("UV",hv(s.uv,1,"")); card("Solare",hv(s.solarWm2,0," W/m2")); card("RSSI",hv(s.rssi,1," dBm")); card("ID ISS",s.locked?String(s.stationId+1U):String("--"));
+  card("Temperatura",hv(s.outTempC,1U," C")); card("Umidita",hv(s.outHumidity,0U," %")); card("Vento",hv(s.windKmh,1U," km/h")); card("Raffica 10m",hv(s.windGustKmh,1U," km/h"));
+  card("Direzione",hv(s.windDirDeg,0U," deg")); card("Pioggia oggi",hv(s.rainDayMm,1U," mm")); card("Rain rate",hv(s.rainRateMmH,1U," mm/h")); card("Pressione",hv(s.pressureHpa,1U," hPa"));
+  card("UV",hv(s.uv,1U,"")); card("Solare",hv(s.solarWm2,0U," W/m2")); card("RSSI",hv(s.rssi,1U," dBm")); card("ID ISS",s.locked?String(s.stationId+1U):String("--"));
   h+=F("</div><h2>Stato</h2><table>");
   h+=F("<tr><td>Rete</td><td>");h+=esc(networkModeName());h+=F(" / ");h+=esc(networkIp());h+=F("</td></tr>");
   h+=F("<tr><td>RF</td><td>");h+=rf.synchronized?F("sincronizzato"):F("ricerca");h+=F(" - ch ");h+=String(rf.channel+1);h+=F(" - ");h+=String(rf.frequencyMhz,6);h+=F(" MHz</td></tr>");
@@ -87,9 +88,9 @@ void initWeb(StationState &station){
     const StationState&s=*statePtr;const DavisRadioStatus&rf=getDavisRadioStatus();String j="{";
     j+="\"firmware\":\""+String(FIRMWARE_VERSION)+"\",\"network\":\""+networkModeName()+"\",\"ip\":\""+networkIp()+"\",";
     j+="\"rf_sync\":"+String(rf.synchronized?"true":"false")+",\"channel\":"+String(rf.channel+1)+",\"frequency_mhz\":"+String(rf.frequencyMhz,6)+",";
-    j+="\"temp_c\":"+jf(s.outTempC)+",\"humidity\":"+jf(s.outHumidity,0)+",\"wind_kmh\":"+jf(s.windKmh)+",\"gust_kmh\":"+jf(s.windGustKmh)+",";
-    j+="\"wind_dir\":"+jf(s.windDirDeg,0)+",\"rain_day_mm\":"+jf(s.rainDayMm)+",\"rain_rate_mmh\":"+jf(s.rainRateMmH)+",\"pressure_hpa\":"+jf(s.pressureHpa)+",";
-    j+="\"uv\":"+jf(s.uv)+",\"solar_wm2\":"+jf(s.solarWm2,0)+",\"rssi\":"+jf(s.rssi)+",\"packets_ok\":"+String(s.packetsOk)+",\"crc_errors\":"+String(s.crcErrors)+"}";
+    j+="\"temp_c\":"+jf(s.outTempC)+",\"humidity\":"+jf(s.outHumidity,0U)+",\"wind_kmh\":"+jf(s.windKmh)+",\"gust_kmh\":"+jf(s.windGustKmh)+",";
+    j+="\"wind_dir\":"+jf(s.windDirDeg,0U)+",\"rain_day_mm\":"+jf(s.rainDayMm)+",\"rain_rate_mmh\":"+jf(s.rainRateMmH)+",\"pressure_hpa\":"+jf(s.pressureHpa)+",";
+    j+="\"uv\":"+jf(s.uv)+",\"solar_wm2\":"+jf(s.solarWm2,0U)+",\"rssi\":"+jf(s.rssi)+",\"packets_ok\":"+String(s.packetsOk)+",\"crc_errors\":"+String(s.crcErrors)+"}";
     server.send(200,"application/json",j);
   });
   server.onNotFound([](){server.send(404,"text/plain","Not found");});
